file cleanup and navbar responsive patch
