--- src/Components/About Me/about.css
@@ -1,12 +1,0 @@
-.test{
-    background-image: url("./blob.svg");
-}
-/* .test2{
-    background-image: url("./blob3.svg");
-    background-image: url("./blob2.svg");
-}
-@media(max-width:1300px){
-    .test2{
-        background-image: none
-    }
-} */
